Container & component directory reorganisation
