+++ app/src/containers/Conversations.vue
@@ -69,6 +69,7 @@
 
 <script setup lang="ts">
 import { ChevronDownIcon } from "@/components/icons";
+import TimelineBlock from "@/components/synergy/timeline/TimelineBlock.vue";
 import { useCommunityService } from "@/composables/useCommunityService";
 import { useModalStore } from "@/stores";
 import { useModel } from "@coasys/ad4m-vue-hooks";
@@ -77,7 +78,6 @@ import { SynergyGroup } from "@coasys/flux-utils";
 import { storeToRefs } from "pinia";
 import { ref, watch } from "vue";
 import { useRoute, useRouter } from "vue-router";
-import TimelineBlock from "../synergy/timeline/TimelineBlock.vue";
 
 interface Props {
   parentChannel: Channel;
