+++ app/src/utils/helperFunctions.ts
